+++ packages/istanbul-widget/src/core/core.ts
@@ -1,11 +1,12 @@
 import { deepMerge, isArray, isFunction, isObject, set } from '@minko-fe/lodash-pro'
+import { type ConsolaInstance, LogLevels, createConsola } from 'consola'
 import { ButtonGroupPlugin } from '@/plugins/button-group/ButtonGroupPlugin'
 import { ReportPlugin } from '@/plugins/report/ReportPlugin'
 import { SettingPlugin } from '@/plugins/setting/SettingPlugin'
 import { ISTANBUL_WIDGET_ID } from '@/utils/const'
 import { $ } from '@/utils/query'
 import Context from './Context'
-import { type IstanbulWidgetOptions, type PluginName } from './options.interface'
+import { type IstanbulWidgetOptions, type PluginName, type PluginType } from './options.interface'
 import { IstanbulWidgetPlugin } from './plugin/IstanbulWidgetPlugin'
 import { IstanbulWidgetReactPlugin } from './plugin/IstanbulWidgetReactPlugin'
 import { type CompInstance, render } from './render'
@@ -41,9 +42,16 @@ export class IstanbulWidget {
 
   public static Context: typeof Context
 
+  public static logger: ConsolaInstance
+
   constructor(opts: IstanbulWidgetOptions) {
+    IstanbulWidget.logger = createConsola({
+      level: opts.debug ? LogLevels.debug : LogLevels.box,
+      fancy: true,
+    })
+
     if (!!IstanbulWidget.instance && IstanbulWidget.instance instanceof IstanbulWidget) {
-      console.debug('[istanbul-widget] IstanbulWidget is already exists.')
+      IstanbulWidget.logger.debug('[istanbul-widget] IstanbulWidget is already exists.')
       return IstanbulWidget.instance
     }
 
@@ -76,16 +84,16 @@ export class IstanbulWidget {
         _onload()
       }
     } else {
-      let _timer
+      let _timer: number
       const _pollingDocument = () => {
         if (!!document && document.readyState === 'complete') {
           _timer && clearTimeout(_timer)
           _onload()
         } else {
-          _timer = setTimeout(_pollingDocument, 1)
+          _timer = window.setTimeout(_pollingDocument, 1)
         }
       }
-      _timer = setTimeout(_pollingDocument, 1)
+      _timer = window.setTimeout(_pollingDocument, 1)
     }
   }
 
@@ -123,7 +131,7 @@ export class IstanbulWidget {
    */
   private _addBuiltInPlugins() {
     // add default report plugin
-    this.addPlugin(new ReportPlugin(`${ISTANBUL_WIDGET_ID}_report__`, '上报插件'))
+    this.addPlugin(new ReportPlugin('report', '上报插件'))
 
     // add other built-in plugins according to user's config
     const list = this.option.defaultPlugins
@@ -149,9 +157,9 @@ export class IstanbulWidget {
       for (let i = 0; i < list.length; i++) {
         const pluginConf = plugins[list[i]]
         if (pluginConf) {
-          this.addPlugin(new pluginConf.proto(`${ISTANBUL_WIDGET_ID}_${list[i]}__`, pluginConf.name, pluginConf.props))
+          this.addPlugin(new pluginConf.proto(list[i], pluginConf.name, pluginConf.props))
         } else {
-          console.debug('[istanbul-widget] Unrecognized default plugin ID:', list[i])
+          IstanbulWidget.logger.debug('[istanbul-widget] Unrecognized default plugin ID:', list[i])
         }
       }
     }
@@ -163,7 +171,7 @@ export class IstanbulWidget {
   public addPlugin(plugin: IstanbulWidgetPlugin) {
     // ignore this plugin if it has already been installed
     if (this.pluginList[plugin.id] !== undefined) {
-      console.debug(`[istanbul-widget] Plugin \`${plugin.id}\` has already been added.`)
+      IstanbulWidget.logger.debug(`[istanbul-widget] Plugin \`${plugin.id}\` has already been added.`)
       return false
     }
     this.pluginList[plugin.id] = plugin
@@ -183,7 +191,8 @@ export class IstanbulWidget {
     set(this.compInstance.pluginList, plugin.id, {
       id: plugin.id,
       name: plugin.name,
-    })
+      domID: plugin.domID,
+    } as PluginType)
 
     this.compInstance.pluginList = this._reorderPluginList(this.compInstance.pluginList)
 
@@ -194,12 +203,17 @@ export class IstanbulWidget {
       })
       .then(() => {
         // start init
-        plugin.event.emit('init')
+        plugin.emit('init')
         // render
-        plugin.event.emit('render', () => {})
+        plugin.emit('render', ({ htmlElement }) => {
+          if (htmlElement) {
+            this.compInstance.pluginList[plugin.id].htmlElement = htmlElement
+            this.compInstance.update({ pluginList: this.compInstance.pluginList })
+          }
+        })
         // end init
         plugin.isReady = true
-        plugin.event.emit('ready')
+        plugin.emit('ready')
       })
   }
 
@@ -211,17 +225,19 @@ export class IstanbulWidget {
       this._initPlugin(this.pluginList[id])
     }
 
-    this.triggerEvent('ready')
+    this.emitEvent('ready')
 
-    if (process.env.NODE_ENV === 'development') {
-      console.log(`[istanbul-widget]: v${this.version}`)
-    }
+    IstanbulWidget.logger.debug(`[istanbul-widget]: v${this.version}`)
   }
 
   /**
-   * Trigger a `istanbulWidget.option` event.
+   * emit a `istanbulWidget.option` event
+   * @example
+   * ```js
+   * istanbulWidget.emitEvent('ready') // will emit `istanbulWidget.option.onReady()`
+   * ```
    */
-  public triggerEvent(eventName: string, param?: any) {
+  public emitEvent(eventName: string, param?: any) {
     eventName = `on${eventName.charAt(0).toUpperCase()}${eventName.slice(1)}`
     if (isFunction(this.option[eventName])) {
       setTimeout(() => {
@@ -275,7 +291,7 @@ export class IstanbulWidget {
 
   public static set instance(value: IstanbulWidget | undefined) {
     if (value !== undefined && !(value instanceof IstanbulWidget)) {
-      console.debug(
+      IstanbulWidget.logger.debug(
         '[istanbul-widget] Cannot set `IstanbulWidget.instance` because the value is not the instance of IstanbulWidget.',
       )
       return