+++ packages/istanbul-widget/src/core/plugin/IstanbulWidgetPlugin.ts
@@ -1,12 +1,12 @@
-import type ReactDOM from 'react-dom/client'
-import { uniqueId } from '@minko-fe/lodash-pro'
-import { Emitter } from 'strict-event-emitter'
+import { isFunction, uniqueId } from '@minko-fe/lodash-pro'
+import EventEmitter from 'eventemitter3'
+import { ISTANBUL_WIDGET_ID } from '@/utils/const'
 import { type IstanbulWidget } from '../core'
 
 type Events = {
   init: []
   ready: []
-  render: [callback?: (res: { root: ReactDOM.Root; domNode: HTMLDivElement }) => void]
+  render: [callback: (res: { htmlElement: HTMLElement | undefined }) => void]
 }
 
 export class IstanbulWidgetPlugin {
@@ -15,12 +15,21 @@ export class IstanbulWidgetPlugin {
   protected _name!: string
   protected _istanbulWidget!: IstanbulWidget
 
-  event: Emitter<Events> = new Emitter()
+  public domID: string
 
-  constructor(id: string, name: string) {
+  private _event: EventEmitter<Events> = new EventEmitter()
+
+  constructor(
+    id: string,
+    name: string,
+    protected htmlElement?: HTMLElement,
+  ) {
     this.id = id
+    this.domID = `${ISTANBUL_WIDGET_ID}__plugin--${id}`
     this.name = name
     this.isReady = false
+
+    this.registerEvents()
   }
 
   get id() {
@@ -33,7 +42,7 @@ export class IstanbulWidgetPlugin {
     } else if (!value) {
       throw '[istanbul-widget] Plugin ID cannot be empty.'
     }
-    this._id = value.toLowerCase()
+    this._id = value
   }
 
   get name() {
@@ -56,6 +65,40 @@ export class IstanbulWidgetPlugin {
       throw '[istanbul-widget] istanbulWidget cannot be empty'
     }
     this._istanbulWidget = value
+  }
+
+  /**
+   * Add a listener for a given event
+   */
+  public on = this._event.on.bind(this._event)
+  /**
+   * Remove an event listener
+   */
+  public off = this._event.off.bind(this._event)
+  /**
+   * Calls each of the listeners registered for a given event
+   */
+  public emit = this._event.emit.bind(this._event)
+
+  protected registerEvents() {
+    const properties = Object.getOwnPropertyNames(IstanbulWidgetPlugin.prototype)
+
+    const onMethods = properties.filter((name) => name.match(/^on[A-Z].*/) && isFunction(this[name]))
+    onMethods.forEach((methodName) => {
+      this[methodName].call(this)
+    })
+  }
+
+  public onReady() {
+    this.on('ready', () => {
+      this.isReady = true
+    })
+  }
+
+  public onRender() {
+    this.on('render', (callback) => {
+      callback({ htmlElement: this.htmlElement })
+    })
   }
 
   protected getUniqueID(prefix: string = '') {